requirements.txt
ollama==0.4.8

langchain==0.3.24
langchain-community==0.3.22
langchain-core==0.3.56
langchain-ollama==0.3.2
langchain-text-splitters==0.3.8

langgraph==0.3.34

pydantic==2.11.3

openpyxl==3.1.5

python-slugify==8.0.4
